:truck: Move event type mapping to enums
